modgui_script: move file removal on end of upgrade device as package in tmp file can be searched by install script

+++ decompressed/gui_file/etc/modgui_scripts/02_specific.sh
@@ -155,8 +155,10 @@ install_specific() {
 }
 
 remove_wizard_5ghz() {
-  logger_command "Removing 5GHz config from wizard..."
-  rm /www/wizard-cards/*wireless_5G*
+  if [ -n "$(find /www/wizard-cards/ -iname '*wireless_5G*')" ]; then
+    logger_command "Removing 5GHz config from wizard..."
+    rm /www/wizard-cards/*wireless_5G*
+  fi
 }
 
 #THIS CHECK DEVICE TYPE AND INSTALL SPECIFIC FILE
